refactor: package.json file changes and unnecessery packages are removed
