+++ package.json
@@ -36,14 +36,13 @@
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "@babel/plugin-transform-modules-commonjs": "^7.15.4",
     "@babel/preset-env": "7.15.8",
-    "@commitlint/cli": "^13.2.1",
-    "@commitlint/config-conventional": "^13.2.0",
     "@storybook/addon-a11y": "^6.4.0",
     "@storybook/addon-actions": "^6.4.0",
     "@storybook/addon-essentials": "^6.4.0",
     "@storybook/react": "^6.4.0",
     "@testing-library/jest-dom": "^5.15.1",
     "@testing-library/react": "^12.1.2",
+    "@testing-library/react-hooks": "^7.0.2",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.0.3",
     "@types/jest-axe": "^3.5.3",
@@ -54,7 +53,6 @@
     "@typescript-eslint/parser": "^4.32.0",
     "babel-loader": "^8.1.0",
     "commitizen": "^4.2.4",
-    "cz-conventional-changelog": "3.3.0",
     "eslint": "^7.32.0",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "husky": "^7.0.4",
@@ -62,8 +60,6 @@
     "jest-axe": "^5.0.1",
     "lerna": "^4.0.0",
     "mkdirp": "^1.0.4",
-    "msw": "^0.35.0",
-    "npm-run-all": "^4.1.5",
     "prettier": "^2.5.0",
     "rimraf": "^3.0.2",
     "scripty": "^2.0.0",
@@ -75,10 +71,8 @@
     "webpack-cli": "^4.9.1"
   },
   "dependencies": {
-    "@testing-library/react-hooks": "^7.0.2",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-remove-scroll": "^2.4.3"
+    "react-dom": "^17.0.2"
   },
   "config": {
     "commitizen": {
