+++ packages/components/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@ciceksepeti/cui",
   "version": "0.0.1",
-  "description": "ÇiçekSepeti Component Button",
+  "description": "ÇiçekSepeti Cactus UI All Packages",
   "author": "cs",
   "main": "dist/src/index.js",
   "license": "MIT",
@@ -10,8 +10,9 @@
     "access": "public"
   },
   "repository": {
-    "type": "",
-    "url": ""
+    "type": "git",
+    "directory": "packages/components",
+    "url": "git+https://github.com/ciceksepetitech/cactus-ui.git"
   },
   "files": [
     "dist",
@@ -30,6 +31,10 @@
     "@ciceksepeti/cui-focus-trap": "^0.0.1",
     "@ciceksepeti/cui-alert-dialog": "^0.0.1",
     "@ciceksepeti/cui-visually-hidden": "^0.0.1"
+  },
+  "peerDependencies": {
+    "react": "^16.8.0 || 17.x",
+    "react-dom": "^16.8.0 || 17.x"
   },
   "scripts": {
     "build": "scripty",
