+++ packages/focus-trap/package.json
@@ -15,12 +15,17 @@
     "README.md"
   ],
   "repository": {
-    "type": "",
-    "url": ""
+    "type": "git",
+    "directory": "packages/focus-trap",
+    "url": "git+https://github.com/ciceksepetitech/cactus-ui.git"
   },
   "dependencies": {
     "@ciceksepeti/cui-utils": "^0.0.1",
     "@ciceksepeti/cui-hooks": "^0.0.1"
+  },
+  "peerDependencies": {
+    "react": "^16.8.0 || 17.x",
+    "react-dom": "^16.8.0 || 17.x"
   },
   "scripts": {
     "build": "scripty",
