+++ packages/hooks/package.json
@@ -18,8 +18,13 @@
     "@ciceksepeti/cui-utils": "^0.0.1"
   },
   "repository": {
-    "type": "",
-    "url": ""
+    "type": "git",
+    "directory": "packages/hooks",
+    "url": "git+https://github.com/ciceksepetitech/cactus-ui.git"
+  },
+  "peerDependencies": {
+    "react": "^16.8.0 || 17.x",
+    "react-dom": "^16.8.0 || 17.x"
   },
   "scripts": {
     "build": "scripty",
