+++ packages/hooks/src/__test__/useArray.test.tsx
@@ -1,5 +1,5 @@
 import { useArray } from '..';
-import { cleanup, reactHooks } from '@ciceksepeti/cui-utils';
+import { cleanup, reactHooks } from '../../../../utils/test-setup';
 
 const { renderHook, act } = reactHooks;
