+++ packages/hooks/src/__test__/useEventListener.test.tsx
@@ -1,7 +1,7 @@
 import React, { useRef } from 'react';
 import { useEventListener } from '..';
 import userEvent from '@testing-library/user-event';
-import { cleanup, screen, render } from '@ciceksepeti/cui-utils';
+import { cleanup, screen, render } from '../../../../utils/test-setup';
 
 const fnMock = jest.fn();
 const fnNullMock = jest.fn();
