+++ packages/hooks/src/__test__/useIsMounted.test.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { useIsMounted } from '..';
-import { cleanup, screen, render } from '@ciceksepeti/cui-utils';
+import { cleanup, screen, render } from '../../../../utils/test-setup';
 
 const fnMock = jest.fn();
 const fnSpecificMock = jest.fn();
