+++ packages/hooks/src/__test__/usePrevious.test.tsx
@@ -1,5 +1,5 @@
 import { usePrevious } from '..';
-import { cleanup, reactHooks } from '@ciceksepeti/cui-utils';
+import { cleanup, reactHooks } from '../../../../utils/test-setup';
 
 const { renderHook } = reactHooks;
