+++ packages/listbox/package.json
@@ -17,13 +17,18 @@
     "styles.css"
   ],
   "repository": {
-    "type": "",
-    "url": ""
+    "type": "git",
+    "directory": "packages/listbox",
+    "url": "git+https://github.com/ciceksepetitech/cactus-ui.git"
   },
   "dependencies": {
     "@ciceksepeti/cui-utils": "^0.0.1",
     "@ciceksepeti/cui-hooks": "^0.0.1",
     "@ciceksepeti/cui-popover": "^0.0.1"
+  },
+  "peerDependencies": {
+    "react": "^16.8.0 || 17.x",
+    "react-dom": "^16.8.0 || 17.x"
   },
   "scripts": {
     "build": "scripty",
