+++ packages/listbox/src/__test__/listbox.test.tsx
@@ -2,7 +2,12 @@ import React from 'react';
 import { axe } from 'jest-axe';
 import Listbox, { ListboxItem } from '..';
 import userEvent from '@testing-library/user-event';
-import { render, screen, cleanup, fireEvent } from '@ciceksepeti/cui-utils';
+import {
+  render,
+  screen,
+  cleanup,
+  fireEvent
+} from '../../../../utils/test-setup';
 
 let consoleWarn;
