+++ packages/portal/package.json
@@ -11,15 +11,17 @@
     "access": "public"
   },
   "repository": {
-    "type": "",
-    "url": ""
+    "type": "git",
+    "directory": "packages/portal",
+    "url": "git+https://github.com/ciceksepetitech/cactus-ui.git"
   },
   "files": [
     "dist",
     "README.md"
   ],
-  "dependencies": {
-    "@ciceksepeti/cui-utils": "^0.0.1"
+  "peerDependencies": {
+    "react": "^16.8.0 || 17.x",
+    "react-dom": "^16.8.0 || 17.x"
   },
   "scripts": {
     "build": "scripty",
