+++ packages/skip-nav/package.json
@@ -17,11 +17,16 @@
     "styles.css"
   ],
   "repository": {
-    "type": "",
-    "url": ""
+    "type": "git",
+    "directory": "packages/skip-nav",
+    "url": "git+https://github.com/ciceksepetitech/cactus-ui.git"
   },
   "dependencies": {
     "@ciceksepeti/cui-utils": "^0.0.1"
+  },
+  "peerDependencies": {
+    "react": "^16.8.0 || 17.x",
+    "react-dom": "^16.8.0 || 17.x"
   },
   "scripts": {
     "build": "scripty",
