+++ packages/utils/package.json
@@ -15,8 +15,13 @@
     "README.md"
   ],
   "repository": {
-    "type": "",
-    "url": ""
+    "type": "git",
+    "directory": "packages/utils",
+    "url": "git+https://github.com/ciceksepetitech/cactus-ui.git"
+  },
+  "peerDependencies": {
+    "react": "^16.8.0 || 17.x",
+    "react-dom": "^16.8.0 || 17.x"
   },
   "scripts": {
     "build": "scripty",
