+++ packages/utils/src/__test__/mergeEventHandlers.test.ts
@@ -1,6 +1,6 @@
 import { SyntheticEvent } from 'react';
 import { mergeEventHandlers } from '..';
-import { cleanup } from '@ciceksepeti/cui-utils';
+import { cleanup } from '../../../../utils/test-setup';
 
 let userMockFn;
 let libraryMockFn;
