+++ packages/utils/src/index.tsx
@@ -1,6 +1,5 @@
 export * from './isCSR';
 export * from './platform';
-export * from './testSetup';
 export * from './polymorphic';
 export * from './conditionalWrapper';
 export * from './setRevertableStyle';
